removed useRemoteProps hook
